TST-8 Moved newsletter-alert.js to the global dir
