+++ assets/js/theme/global.js
@@ -19,7 +19,7 @@ import associateAffiliate from './global/associate-affiliate';
 import joinProcessInteraction from './global/joinProcessInteraction';
 import shippingModalInteraction from './global/shippingModalInteraction';
 import accordian from './global/accordian';
-import newsletterAlert from './razoyo/newsletter-alert';
+import newsletterAlert from './global/newsletter-alert';
 
 
 export default class Global extends PageManager {
